Moved requireFromPath to util folder
