+++ src/script/modules/misc.ts
@@ -1,6 +1,6 @@
 import { pumlhorse } from '../../PumlhorseGlobal';
 
-var requireFromPath = require('../../util/requireFromPath');
+var requireFromPath = require('../../../util/requireFromPath');
 
 function importFunc(moduleName) {
     return requireFromPath(moduleName, this.__filename);
