+++ tsconfig.json
@@ -5,8 +5,5 @@
     },
     "include": [
         "./src/**/*"
-    ],
-    "exclude": [
-        "./src/util/requireFromPath.js"
     ]
 }
